chore(scenario): move i18n type to file
