+++ packages/types/src/index.ts
@@ -2,3 +2,4 @@ export * from './request';
 export * from './response';
 export * from './salute';
 export * from './global';
+export * from './i18n';
