+++ packages/types/src/salute.ts
@@ -2,6 +2,7 @@
 import { AppInfo, CharacterId } from './global';
 import { AppState, Message, NLPRequest } from './request';
 import { NLPResponse, ErrorCommand, EmotionType, Button } from './response';
+import { KeysetDictionary, I18nOptions } from './i18n';
 
 interface IntentSlot {
     name: string; // имя сущности
@@ -129,18 +130,3 @@ export type ScenarioSchema = Record<
         children?: ScenarioSchema;
     }
 >;
-
-export interface IPluralForms {
-    one: string;
-    some: string;
-    many?: string;
-    none?: string;
-}
-export type I18nBaseOptions = Record<string, string | Record<string, unknown> | number | undefined>;
-export type I18nPluralOptions = I18nBaseOptions & {
-    count: number;
-};
-export type I18nOptions = I18nBaseOptions | I18nPluralOptions;
-export type KeysetKey = string | IPluralForms;
-export type Keyset = Record<string, KeysetKey>;
-export type KeysetDictionary = Record<string, Keyset>;
